Update message errors when id not found

+++ server/src/resources/catalog/catalog.controllers.js
@@ -7,13 +7,16 @@ const findMany = async (req,res)=>{
         res.status(200).json({results: docs});
     }catch(e){
         console.log(e);
-        res.status(500).json({error:'Internal error'});
+        res.status(500).json({error:'Internal Server error'});
     }
 }
 
 const createOne = async (req,res)=>{
     try{
         const newCatalog = req.body;
+        if(JSON.stringify(newCatalog)=='{}'){
+            return res.status(400).json({error:'Creation failed because there are no input data'});
+        }
         const doc = await Catalog.create(newCatalog);
         res.status(201).json({results: doc})
     }catch(e){
@@ -23,25 +26,29 @@ const createOne = async (req,res)=>{
 }
 
 const findOne = async (req, res) =>{
+    const {id} = req.params;
     try{
-        const {id} = req.params;
-        const doc = await Catalog.find({_id: id});
+        const doc = await Catalog.findOne({_id: id});
         if(!doc){
-            return res.status(404).json({error:'Not found'});
+            return res.status(404).json({error:`Catalog with ID ${id} not found.`});
         }
         res.status(200).json({results:doc});
     }catch(e){
         console.log(e);
-        res.status(500).json({error:'Internal error'});
+        res.status(500).json({error:'Internal Server error'});
     }
 }
 
 const updateOne = async (req,res)=>{
+    const {id} = req.params;
     try{
-        const {id} = req.params;
-        const doc = await Catalog.findOneAndUpdate({_id: id},req.body,{new:true});//"new: true" to return the document after update
+        const updatedCatalog = req.body;
+        if(JSON.stringify(updatedCatalog)=='{}'){
+            return res.status(400).json({error:'Updated failed because there are no input data'});
+        }
+        const doc = await Catalog.findOneAndUpdate({_id: id},updatedCatalog,{new:true});
         if(!doc){
-            return res.status(404).json({error:'Not found'});
+            return res.status(404).json({error:`Catalog with ID ${id} not found.`});
         }
         res.status(200).json({results:doc});
     }catch(e){
@@ -51,11 +58,11 @@ const updateOne = async (req,res)=>{
 }
 
 const deleteOne = async (req,res)=>{
+    const {id} = req.params;
     try{
-        const {id} = req.params;
         const doc = await Catalog.findOneAndDelete({_id: id}); 
         if(!doc){
-            return res.status(404).json({error:'Not found'});
+            return res.status(404).json({error:`Catalog with ID ${id} not found.`});
         }
         res.status(200).json({results:doc});
     }catch(e){
@@ -65,11 +72,11 @@ const deleteOne = async (req,res)=>{
 }
 
 const findProductsById = async (req,res)=>{
+    const {catalogid} = req.params;
     try{
-        const {catalogid} = req.params;
         const docs = await Product.find({catalog_id: catalogid});
-        if(docs === null){
-            res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
+        if(!docs){
+            return res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
         }
         res.status(200).json({results: docs});
     }catch(e){
@@ -79,14 +86,21 @@ const findProductsById = async (req,res)=>{
 }
 
 const createProductById = async (req,res)=>{
+    const {catalogid} = req.params;
     try{
         const newProduct = req.body;
-        const {catalogid} = req.params;
+        //comprovar que hi ha JSON d'entrada
+        if(JSON.stringify(newProduct)=='{}'){  
+            return res.status(400).json({error:'Updated failed because there are no input data'});
+        }
+        //comprovar que el catalog_id existeix
+        const catalogDoc = await Catalog.findOne({_id:catalogid});
+        if(!catalogDoc){
+            return res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
+        }
+        //afegir el producte al cataleg
         newProduct.catalog_id = catalogid;
         const doc = await Product.create(newProduct);
-        if(doc === null){
-            res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
-        }
         res.status(201).json({results: doc});
     }catch(e){
         console.log(e);
@@ -95,14 +109,20 @@ const createProductById = async (req,res)=>{
 }
 
 const updateProductById = async (req,res)=>{
+    const {id,catalogid} = req.params;
     try{
         const updatedProduct = req.body;
-        const {id,catalogid} = req.params;
+        if(JSON.stringify(updatedProduct)=='{}'){
+            return res.status(400).json({error:'Updated failed because there are no input data'});
+        }
+        const catalogDoc = await Catalog.findOne({_id:catalogid});
+        if(!catalogDoc){
+            return res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
+        }
         updatedProduct.catalog_id = catalogid;
         const doc = await Product.findOneAndUpdate({_id: id},updatedProduct,{new:true});
-        Catalog.findOneAndUpdate({_id: id},req.body,{new:true});
-        if(doc === null){
-            res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
+        if(!doc){
+            return res.status(400).send({error: `Product with ID ${id} not found.`})
         }
         res.status(201).json({results: doc});
     }catch(e){
@@ -112,11 +132,15 @@ const updateProductById = async (req,res)=>{
 }
 
 const deleteProductById = async (req,res)=>{
+    const {id, catalogid} = req.params;
     try{
-        const {id, catalogid} = req.params;
+        const catalogDoc = await Catalog.findOne({_id:catalogid});
+        if(!catalogDoc){
+            return res.status(400).send({error: `Catalog with ID ${catalogid} not found.`})
+        }
         const doc = await Product.findOneAndDelete({$and: [{_id: id, catalog_id: catalogid}]});
         if(!doc){
-            return res.status(404).json({error:'Not found'});
+            return res.status(404).json({error:`Product with ID ${id} not found.`});
         }
         res.status(200).json({results: doc});
     }catch(e){
